requirements.txt
# Core Flask & Web Framework
Flask==2.3.1
flask_sqlalchemy==3.0.3
flask_cors==4.0.0
Werkzeug==2.3.1

# Task Queue & Background Processing
celery==5.3.1
redis==5.0.0

# AI/ML & Language Models
langchain>=0.3.0
langchain-core>=0.3.55
langchain-community>=0.3.0
langchain-openai>=0.3.14
langchain-text-splitters>=0.1.0
openai>=1.68.2
tiktoken==0.7.0

# Vector Store & Embeddings
pinecone-client>=6.0.0
langchain-pinecone>=0.2.4
faiss-cpu==1.10.0

# PDF Processing & Document Handling
pypdf==3.15.4
PyMuPDF==1.25.5

# Computer Vision & OCR
opencv-python>=4.8.0
pytesseract>=0.3.10
pymupdf>=1.22.5
Pillow>=10.0.0
numpy>=1.24.0

# AWS & Cloud Services
boto3==1.26.122
botocore==1.29.122

# Observability & Monitoring
langfuse==2.57.13

# Development & Utilities
invoke==2.2.0
python-dotenv==1.0.0
watchdog==3.0.0
click==8.1.3
backoff==2.2.1
uuid==1.30
